// types/CalPurOrdInput.ts
export interface CalPurOrdInput {
  cmdGrp3: CommandGroup3
}
export interface CommandGroup3 {
  tPurOrd?: CommandGroup3TPurOrd[]
}
export interface CommandGroup3TPurOrd {
  poId?: number
  poNo?: string
  poDate?: string
  saleJobDesc?: string
  supId: number
  supName?: string
  paidSupId?: number
  mDocPrefixRunProdClassId?: number
  mSupProdClassId?: number
  onePoFlag?: string
  maxPriceFlag?: string
  paidSupName?: string
  imptFlag?: string
  poRefTypIdx?: number
  poRefId?: number
  saleJobId?: number
  othRefNo?: string
  changeOrdQtyFlag?: string
  supRefNo?: string
  cashCrTerm?: number
  crTerm?: number
  expireDate?: string
  tradeIncoTermTyp?: string
  paymentTypIdx?: number
  supPaymentDueTypIdx?: number
  balCrLineAmt?: number
  paymentTermIdx?: number
  pmPersonId?: number
  cncyId?: number
  presentVatRptFlag?: string
  exchgRate?: number
  vatRptFlag?: string
  poAmt?: number
  reqDate?: string
  purLeadtime?: number
  taxNo?: string
  headOfficeFlag?: string
  poExpireDay?: number
  supBranchCode?: string
  reqDateTypIdx?: number
  supBillName: string
  supBillNameItem?: number
  supBillAddr: string
  supBillAddrItem?: number
  bankBasketId?: number
  apAdvNoInvBfAmt?: number
  focFlag?: string
  poRcvFlag?: string
  discPct1?: number
  discPct2?: number
  discPct3?: number
  cashDiscPct?: number
  discAmt?: number
  prodClassProdTyp?: string
  chngPriceFlag?: string
  holdApvTimeIdx?: number
  specificPurProdTypIdx?: number
  vatInclFlag?: string
  vatPct?: number
  poBeforeVatAmt?: number
  vatAmt?: number
  advPctFlag?: string
  advPoPct?: number
  advPoAmt?: number
  adjBeforeVatAmt?: number
  adjBeforePoAmt?: number
  defSupInvDepId?: number
  defSupInvDepName?: string
  defRcvDepId?: number
  defRcvDepName?: string
  purShipTypIdx?: number
  rcvDepAddrItem?: number
  styleId?: number
  assortId?: number
  assortSolidTypIdx?: number
  noRtnFlag?: string
  supShipAddr?: string
  supShipAddrItem?: number
  firstShipDate?: string
  shipMinPct?: number
  sumVatAbleAmt?: number
  sumNonVatAbleAmt?: number
  shipExtraPct?: number
  partialShipFlag?: string
  xchngFlag?: string
  platformIdx?: number
  platformLockFlag?: string
  billPaymentRef1?: string
  billPaymentRef2?: string
  billPaymentRef3?: string
  manualCloseQRFlag?: string
  manualCloseQRGrpPict?: number
  stkFilterFlag?: string
  stkTypIdx?: number
  cUsrName?: string
  lUsrName?: string
  minReqDate?: string
  sumVatBaseAbleAmt?: number
  maxPurLeadtime?: number
  cfmUsrName?: string
  docPrefixRunId?: number
  remark?: string
  sumPoAmt?: number
  sumPoVatAmt?: number
  cObjId?: number
  createTyp?: string
  prodTypId?: number
  comId: number
  depId: number
  cfmTyp: string
  tPoCoupon?: CommandGroup3TPoCoupon[]
  tPoProd: CommandGroup3TPoProd[]
}
export interface CommandGroup3TPoCoupon {
  platformCouponId?: number
  poId?: number
  xsysId?: number
  prodId?: number
  prodName?: string
  uomId?: number
  poQty?: number
  brandId?: number
  poAmt?: number
  discAmt?: number
  poItem: number
}
export interface CommandGroup3TPoProd {
  poId?: number
  poItem?: number
  xsysId?: number
  itemNo?: number
  prodId: number
  supProdName: string
  assortId?: number
  focFlag?: string
  stkQty?: number
  dozQty?: number
  assortQty?: number
  calQtyFormula?: string
  poQty?: number
  supUomId: number
  prodWgt?: number
  lastPurQty?: number
  lastPurDateDesc?: string
  lastPurPriceDesc?: string
  labelPrice?: number
  calPriceFormula?: string
  priceUomId: number
  perUomVlm?: number
  perUomWgt?: number
  pricePoQty?: number
  unitPrice?: number
  discPct1?: number
  discPct2?: number
  discPct3?: number
  discPerUom?: number
  poAmt?: number
  discAmt?: number
  itemProdFlag?: string
  fixPriceRatioFlag?: string
  fixRatioFlag?: string
  lotTypIdx?: number
  netPurPrice?: number
  focRemark?: string
  stdNetPurPrice?: number
  supUnitPrice?: number
  prodVlm?: number
  prodLotOrdFlag?: string
  platformGpPct?: number
  remark?: string
  poRefItem?: number
  poNetAmt?: number
  poVatAmt?: number
  vatAbleAmt?: number
  vatTypIdx?: number
  nonVatAbleAmt?: number
  transBeforeAdvAmt?: number
  vatBeforeAdvAmt?: number
  prodGrpPict?: number
  prodBarcode?: string
  prodName?: string
  headDiscAmt?: number
  vatBaseAmt?: number
  vatBaseAbleAmt?: number
}